Reuse the RFC7231InstantFormatter for all Netty Channels
Problem

A new DateTimeFormatter is created for each Netty Channel. In the HTTP/2
flow this can start to add up as each dispatch will get its own Channel.

Solution

Create it once in the object and share it. The DateTimeFormatter type is
immutable and thread safe so there is no reason not to.