Use the Netty EventLoop for execution in the Http4sNettyHandler (#394)
* Use the Netty EventLoop for execution in the Http4sNettyHandler

Problem

The Netty pipelines expect you to be in the EventLoop thread associated
with that pipeline when you perform pipeline events such as calling
ctx.read(), ctx.write(..), etc. If you're not, events are shifted on
your behalf into the channels event loop. This shifting can make it tough
to reason about when events will occur, and in what order.

The comment about the Trampoline enforcing write orders is also a bit
misleading: the trampoline is a thread-local construct so it doesn't
enforce any ordering between different threads. It is the sequencing of
the Futures that is really enforcing the desired ordering: the
Future[(Response,Cleanup)], while it may be ready in any order, is not
acted upon (via the contents in .map {..}) until the previous has
finished and thus let the flatMap operation proceed.

Solution

Instead of using a Trampoline ExecutionContext we instead just use the
Netty EventLoop directly. This gives us all the same properties we wanted,
namely serial execution of events, but also makes our Netty events
immediately traverse the pipeline.

* deprecate Trampoline