Add support for server-side request cancellation
Problem

Performing some server computations may be pretty expensive and in the
case where the channel gets shutdown, either because the client hung up
or the H2 stream was cancelled (due to a timeout, for example), the
server will continue that processing in vein.

Solution

Propagate cancellation into the service handling logic in case it can
save itself some work or potentially notify its own backends.