Add support for HTTP/2 Prior Knowledge to the server implementation
Problem

Right now we support HTTP/2 only as a negotiated protocol via ALPN. This
is unfortunate as HTTP/2 is just as good cleartext as it is over TLS.

Solution

Add HTTP/2 prior knowledge detection to the cleartext pathway for the
server. This is done by looking for the magic preface which is not valid
as a HTTP/1.x message preface and initialize the pipeline accordingly.