Fix a race condition
We check if the channel is closed before we submit the dispatch to the
channels executor, so if the channel is initially active but closes before
we enter the executor we will put a request in the queue and it will sit
there forever.

Instead, check that the channel is active right before dispatching. This
closes the window for the race condition.